+++ 14503.h
@@ -1,19 +1,19 @@
-//#pragma once
-//#include"global.h"
-//
-//class Algo
-//{
-//	int dir_;
-//	int x_;
-//	int y_;
-//public:
-//	struct coord {
-//		int x;
-//		int y;
-//		int dir;
-//	};
-//	Algo(int& x, int& y, int& dir);
-//	vector<vector<int>> makeMap(int n, int m);
-//	coord clean(int x, int y, int dir);
-//	coord move(int x, int y, int dir, vector<vector<int>>& adj);
-//};
+#pragma once
+#include"global.h"
+
+class Algo
+{
+	int dir_;
+	int x_;
+	int y_;
+public:
+	struct coord {
+		int x;
+		int y;
+		int dir;
+	};
+	Algo(int& x, int& y, int& dir);
+	vector<vector<int>> makeMap(int n, int m);
+	coord clean(int x, int y, int dir);
+	coord move(int x, int y, int dir, vector<vector<int>>& adj);
+};
