+++ global.h
@@ -1,9 +1,9 @@
-//#pragma once
-//#include<iostream>
-//#include<vector>
-//#include<deque>
-//#include<algorithm>
-//#include<queue>
-//#include<stack>
-//
-//using namespace std;
+#pragma once
+#include<iostream>
+#include<vector>
+#include<deque>
+#include<algorithm>
+#include<queue>
+#include<stack>
+
+using namespace std;
